category_id row added, cosmetics

+++ bookkeeper/dirs/models_dir/expense.py
@@ -6,7 +6,7 @@ from dataclasses import dataclass, field
 class Expense(db.Entity):
     id = PrimaryKey(int, auto = True)
     date = Required(datetime.datetime, default = datetime.datetime.now())
-    amount = Required(int)
+    amount = Required(float)
     category = Required('Category')
     comment = Optional(str)
